+++ include/seqan3/search/fm_index/bi_fm_index.hpp
@@ -72,7 +72,12 @@ private:
     using sdsl_index_type = sdsl_index_type_;
 
     //!\brief The type of the underlying SDSL index for the reversed text.
-    using rev_sdsl_index_type = sdsl_index_type_;
+    using rev_sdsl_index_type = sdsl::csa_wt<sdsl_wt_index_type::wavelet_tree_type, // Wavelet tree type
+                                             10'000'000, // Sampling rate of the suffix array
+                                             10'000'000, // Sampling rate of the inverse suffix array
+                                             sdsl::sa_order_sa_sampling<>, // Text or SA based sampling for SA
+                                             sdsl::isa_sampling<>, // Text or ISA based sampling for ISA
+                                             sdsl_wt_index_type::alphabet_type>; // How to represent the alphabet
 
     /*!\brief The type of the reduced alphabet type. (The reduced alphabet might be smaller than the original alphabet
      *        in case not all possible characters occur in the indexed text.)
@@ -85,8 +90,8 @@ private:
     //!\brief The type of the underlying FM index for the original text.
     using fm_index_type = fm_index<alphabet_t, text_layout_mode_, sdsl_index_type>;
 
-    //!\brief The type of the underlying FM index for the reversed text.\if DEV \todo Change sampling behaviour. \endif
-    using rev_fm_index_type = fm_index<alphabet_t, text_layout_mode_, sdsl_index_type>;
+    //!\brief The type of the underlying FM index for the reversed text.
+    using rev_fm_index_type = detail::reverse_fm_index<alphabet_t, text_layout_mode_, rev_sdsl_index_type>;
     //!\}
 
     //!\brief Underlying FM index for the original text.
@@ -115,49 +120,12 @@ private:
      * No guarantee. \if DEV \todo Ensure strong exception guarantee. \endif
      */
     template <std::ranges::range text_t>
-    //!\cond
-        requires (text_layout_mode_ == text_layout::single)
-    //!\endcond
     void construct(text_t && text)
     {
-        static_assert(std::ranges::bidirectional_range<text_t>, "The text must model bidirectional_range.");
-        static_assert(alphabet_size<range_innermost_value_t<text_t>> <= 256, "The alphabet is too big.");
-        static_assert(std::convertible_to<range_innermost_value_t<text_t>, alphabet_t>,
-                     "The alphabet of the text collection must be convertible to the alphabet of the index.");
-        static_assert(range_dimension_v<text_t> == 1, "The input cannot be a text collection.");
-
-        // text must not be empty
-        if (std::ranges::begin(text) == std::ranges::end(text))
-            throw std::invalid_argument("The text that is indexed cannot be empty.");
-
-        auto rev_text = std::views::reverse(text);
-        fwd_fm = fm_index_type{text};
-        rev_fm = fm_index_type{rev_text};
-    }
-
-    //!\overload
-    template <std::ranges::range text_t>
-    //!\cond
-        requires (text_layout_mode_ == text_layout::collection)
-    //!\endcond
-    void construct(text_t && text)
-    {
-        static_assert(std::ranges::bidirectional_range<text_t>, "The text must model bidirectional_range.");
-        static_assert(std::ranges::bidirectional_range<std::ranges::range_reference_t<text_t>>,
-                      "The elements of the text collection must model bidirectional_range.");
-        static_assert(alphabet_size<range_innermost_value_t<text_t>> <= 256, "The alphabet is too big.");
-        static_assert(std::convertible_to<range_innermost_value_t<text_t>, alphabet_t>,
-                     "The alphabet of the text collection must be convertible to the alphabet of the index.");
-        static_assert(range_dimension_v<text_t> == 2, "The input must be a text collection.");
-
-        // text must not be empty
-        if (std::ranges::begin(text) == std::ranges::end(text))
-            throw std::invalid_argument("The text that is indexed cannot be empty.");
-
-        auto rev_text = text | views::deep{std::views::reverse} | std::views::reverse;
+        detail::fm_index_validator::validate<alphabet_t, text_layout_mode_>(text);
 
         fwd_fm = fm_index_type{text};
-        rev_fm = fm_index_type{rev_text};
+        rev_fm = rev_fm_index_type{text};
     }
 
 public:
@@ -180,8 +148,6 @@ public:
     using cursor_type = bi_fm_index_cursor<bi_fm_index>;
     //!\brief The type of the unidirectional cursor on the original text.
     using fwd_cursor_type = fm_index_cursor<fm_index_type>;
-    //!\brief The type of the unidirectional cursor on the reversed text.
-    using rev_cursor_type = fm_index_cursor<rev_fm_index_type>;
 
     //!\}
 
@@ -308,28 +274,9 @@ public:
      *
      * No-throw guarantee.
      */
-    fwd_cursor_type fwd_begin() const noexcept
+    fwd_cursor_type fwd_cursor() const noexcept
     {
        return {fwd_fm};
-    }
-
-    /*!\brief Returns a unidirectional seqan3::fm_index_cursor on the reversed text of the bidirectional index that
-     *        can be used for searching. Note that because of the text being reversed, extend_right() resp. cycle_back()
-     *        correspond to extend_left() resp. cycle_front() on the bidirectional index cursor.
-     * \attention For text collections the text IDs are also reversed.
-     * \returns Returns a unidirectional seqan3::fm_index_cursor on the index of the reversed text.
-     *
-     * ### Complexity
-     *
-     * Constant.
-     *
-     * ### Exceptions
-     *
-     * No-throw guarantee.
-     */
-    rev_cursor_type rev_begin() const noexcept
-    {
-       return {rev_fm};
     }
 
     /*!\cond DEV