+++ CHANGELOG.md
@@ -120,6 +120,8 @@ Note that 3.1.0 will be the first API stable release and interfaces in this rele
   * The configuration element `seqan3::search_cfg::mode` does not exist anymore.
     You can replace it by directly using one of the above mentioned "hit strategy" configuration elements
     ([\#1639](https://github.com/seqan/seqan3/pull/1639)).
+  * Removed `seqan3::bi_fm_index_cursor::to_rev_cursor()` and `seqan3::bi_fm_index::rev_cursor()`
+    ([\#1892](https://github.com/seqan/seqan3/pull/1892)).
 
 ## Notable Bug-fixes
 
@@ -157,6 +159,8 @@ Note that 3.1.0 will be the first API stable release and interfaces in this rele
 * The `seqan3::fm_index_cursor::extend_right()`, `seqan3::bi_fm_index_cursor::extend_right()` and
   `seqan3::bi_fm_index_cursor::extend_left()` functions handle c-style strings without including the null character
   ([\#1588](https://github.com/seqan/seqan3/pull/1588)).
+* `seqan3::fm_index` and `seqan3::bi_fm_index` construct the index correctly if a collection with a single text is
+  passed as input ([\#1892](https://github.com/seqan/seqan3/pull/1892)).
 
 #### Range